requirements.txt
blinker==1.9.0
click==8.3.1
ecdsa==0.19.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyasn1==0.6.2
python-jose==3.5.0
python-pam==2.0.2
rsa==4.9.1
six==1.17.0
Werkzeug==3.1.6
